refactor(cypress): update import paths for SEO constants in e2e tests
Simplified import paths in Cypress e2e test files by removing the [locale] segment and using the @/ alias for consistency, likely due to directory structure refactoring.
