+++ apps/filecoin-site/cypress/e2e/case_studies.cy.ts
@@ -3,7 +3,7 @@ import { tests } from '@filecoin-foundation/cypress/support'
 import { PATHS } from '@/constants/paths'
 import { BASE_URL } from '@/constants/siteMetadata'
 
-import { CASE_STUDIES_SEO } from '@/[locale]/case-studies/constants/seo'
+import { CASE_STUDIES_SEO } from '@/case-studies/constants/seo'
 
 const { path } = PATHS.CASE_STUDIES
 const { title, description } = CASE_STUDIES_SEO
