+++ apps/filecoin-site/cypress/e2e/provide_storage_onboarding.cy.ts
@@ -3,7 +3,7 @@ import { tests } from '@filecoin-foundation/cypress/support'
 import { PATHS } from '@/constants/paths'
 import { BASE_URL } from '@/constants/siteMetadata'
 
-import { PROVIDE_STORAGE_ONBOARDING_SEO } from '@/[locale]/provide-storage/onboarding/constants/seo'
+import { PROVIDE_STORAGE_ONBOARDING_SEO } from '@/provide-storage/onboarding/constants/seo'
 
 const { path } = PATHS.PROVIDE_STORAGE_ONBOARDING
 const { title, description } = PROVIDE_STORAGE_ONBOARDING_SEO
