+++ apps/filecoin-site/cypress/e2e/store_data.cy.ts
@@ -3,7 +3,7 @@ import { tests } from '@filecoin-foundation/cypress/support'
 import { PATHS } from '@/constants/paths'
 import { BASE_URL } from '@/constants/siteMetadata'
 
-import { STORE_DATA_SEO } from '@/[locale]/store-data/constants/seo'
+import { STORE_DATA_SEO } from '@/store-data/constants/seo'
 
 const { path } = PATHS.STORE_DATA
 const { title, description } = STORE_DATA_SEO
