+++ apps/filecoin-site/src/app/[locale]/blog/[slug]/components/BlogPostHeader.tsx
@@ -9,10 +9,9 @@ import { BASE_URL } from '@/constants/siteMetadata'
 import { Heading } from '@/components/Heading'
 import { SectionDivider } from '@/components/SectionDivider'
 
-
-import { Categories } from '../../components/Categories'
-import { PostMetadata } from '../../components/PostMetadata'
-import type { BlogPost } from '../../types/blogPostType'
+import { Categories } from '@/blog/components/Categories'
+import { PostMetadata } from '@/blog/components/PostMetadata'
+import type { BlogPost } from '@/blog/types/blogPostType'
 
 type BlogPostHeaderProps = {
   categories: BlogPost['categories']
